Day_17|| Role Base Or Resource based authorization || Update or Delete API work Role base

+++ Restaurants.Application/Dishes/Commands/DeleteDishes/DeleteDishesForRestaurantCommandHandler.cs
@@ -1,13 +1,15 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Contanst;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositor;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDishes
 {
     public class DeleteDishesForRestaurantCommandHandler(ILogger<DeleteDishesForRestaurantCommandHandler> logger,
-        IRestaurantsRepository restaurantsRepository,
+        IRestaurantsRepository restaurantsRepository, IRestaurantAuthorizationServicce restaurantAuthorizationServicce,
         IDishesRepository dishesRepository)
         : IRequestHandler<DeleteDishesForRestaurantCommand>
     {
@@ -20,6 +22,8 @@ namespace Restaurants.Application.Dishes.Commands.DeleteDishes
             {
                 throw new NotFoundException($"Restaurant with  Id {request.RestaurantId} found in DB but no any Dish is there to delete.");
             }
+            if (!restaurantAuthorizationServicce.Authorize(restaurant, RessourceOption.Delete))
+                throw new ForbidException();
             await dishesRepository.DeleteAsync(restaurant.Dishes);
         }
     }
